backend/requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile C:/Users/colin/Documents/waterer/backend/requirements/dev.in
#
appdirs==1.4.4
    # via
    #   black
    #   virtualenv
atomicwrites==1.4.0
    # via pytest
attrs==21.2.0
    # via
    #   -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
    #   jsonschema
    #   pytest
bidict==0.21.2
    # via
    #   -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
    #   python-socketio
black==21.5b2
    # via -r C:/Users/colin/Documents/waterer/backend/requirements/dev.in
cfgv==3.3.0
    # via pre-commit
click==8.0.1
    # via
    #   -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
    #   black
    #   flask
colorama==0.4.4
    # via
    #   -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
    #   click
    #   pytest
cycler==0.10.0
    # via matplotlib
distlib==0.3.2
    # via virtualenv
filelock==3.0.12
    # via virtualenv
flask-cors==3.0.10
    # via -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
flask==2.0.1
    # via
    #   -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
    #   flask-cors
identify==2.2.9
    # via pre-commit
iniconfig==1.1.1
    # via pytest
isort==5.8.0
    # via -r C:/Users/colin/Documents/waterer/backend/requirements/dev.in
itsdangerous==2.0.1
    # via
    #   -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
    #   flask
jinja2==3.0.1
    # via
    #   -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
    #   flask
jsonschema==3.2.0
    # via -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
kiwisolver==1.3.1
    # via matplotlib
markupsafe==2.0.1
    # via
    #   -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
    #   jinja2
matplotlib==3.3.4
    # via -r C:/Users/colin/Documents/waterer/backend/requirements/dev.in
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
numpy==1.21.0
    # via
    #   -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
    #   matplotlib
packaging==20.9
    # via pytest
pathspec==0.8.1
    # via black
pillow==8.2.0
    # via matplotlib
pluggy==0.13.1
    # via pytest
pre-commit==2.13.0
    # via -r C:/Users/colin/Documents/waterer/backend/requirements/dev.in
py==1.10.0
    # via pytest
pydantic==1.8.2
    # via -r C:/Users/colin/Documents/waterer/backend/requirements/dev.in
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.17.3
    # via
    #   -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
    #   jsonschema
pyserial==3.5
    # via -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
pytest-env==0.6.2
    # via -r C:/Users/colin/Documents/waterer/backend/requirements/dev.in
pytest==6.2.4
    # via
    #   -r C:/Users/colin/Documents/waterer/backend/requirements/dev.in
    #   pytest-env
python-dateutil==2.8.1
    # via matplotlib
python-engineio==4.2.0
    # via
    #   -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
    #   python-socketio
python-socketio==5.3.0
    # via -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
pyyaml==5.4.1
    # via pre-commit
regex==2021.4.4
    # via black
six==1.16.0
    # via
    #   -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
    #   cycler
    #   flask-cors
    #   jsonschema
    #   python-dateutil
    #   virtualenv
toml==0.10.2
    # via
    #   black
    #   pre-commit
    #   pytest
typing-extensions==3.10.0.0
    # via pydantic
virtualenv==20.4.7
    # via pre-commit
werkzeug==2.0.1
    # via
    #   -r C:/Users/colin/Documents/waterer/backend/requirements\base.txt
    #   flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
